requirements.txt
paramiko==3.4.0
pywinrm==0.4.3
httpx==0.27.2
pydantic==2.9.2
